move seller to page folder from components
